% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateids.R
\name{annotateids}
\alias{annotateids}
\title{Annotate biological IDs with AnnotationDbi::select}
\usage{
annotateids(db, ids, idtype, columns, keep = FALSE, allowable = FALSE)
}
\arguments{
\item{db}{An AnnotationDb object (i.e org.Hs.eg.db)}

\item{ids}{Input IDs (i.e. TP53)}

\item{idtype}{The type of input IDs (i.e. SYMBOL)}

\item{columns}{The output annotation fields that you want (i.e. ENSEMBL)}

\item{keep}{Keep all multiple values returned by select() (default: FALSE)}

\item{allowable}{Show the columns that are available for the database object
as input and output (default: FALSE)}
}
\value{
A data.frame in the same order and length as your input ids (but see
'keep' param)
}
\description{
This function annotates biological identifiers such as Symbols, ENSEMBLIDs, and more. It requires the appropriate AnnotationDb object (OrgDB, EnsDb) that contains your identifiers of interest. Then you choose what fields (i.e. columns) to annotate your input IDs.
}
\examples{
# Using OrgDb database
library(org.Hs.eg.db)

# Show allowable columns
annotateids(org.Hs.eg.db, allowable = TRUE)

# Keep all multi values. Notice that multiple rows are returned for AKAP17A
ids = c("TP53", "AKAP17A")
idtype = "SYMBOL"
annotateids(org.Hs.eg.db, ids = ids, idtype = idtype,
            columns = c("ENSEMBL","GENENAME","GENETYPE"), keep = TRUE)

# Default action is to remove multi values
annotateids(org.Hs.eg.db, ids = ids, idtype = idtype,
            columns = c("ENSEMBL","GENENAME","GENETYPE"))

\dontrun{
# Using EnsDb database (EnsDb object) from AnnotationHub
ah = AnnotationHub::AnnotationHub()
query(ah, c("EnsDb", "Homo sapiens"))  # accession AH116291 is version 111
edb = query(ah, pattern = c("Homo sapiens", "EnsDb", 111))[[1]]
ids = c("ENSG00000141510", "ENSG00000236362")

annotateids(edb, ids = ids, idtype = "GENEID",
            columns = c("SYMBOL","DESCRIPTION","GENEBIOTYPE"))
}
}
