% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_info.R
\name{object_info}
\alias{object_info}
\title{Information about an R object}
\usage{
object_info(obj)
}
\arguments{
\item{obj}{An R object}
}
\value{
Nothing
}
\description{
Get information about an R object such as class, typeof, etc...
}
\examples{
object_info("hello")
object_info(mtcars)
}
