% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{Peek inside an object}
\usage{
peek(data, row = 5, col = 5, glimpse = FALSE)
}
\arguments{
\item{data}{An object}

\item{row}{Number of rows to show}

\item{col}{Number of columns to show}

\item{glimpse}{Run dplyr::glimpse?}
}
\value{
Nothing
}
\description{
Show the first few rows and columns of an object
}
\examples{
peek(mtcars, glimpse = TRUE)
peek(letters)
}
