% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_operations.R
\name{set_operations}
\alias{set_operations}
\title{Perform Set Operations on Two Elements of a List}
\usage{
set_operations(x, category_names = NULL, max_show = 10, max_setdiff_show = Inf)
}
\arguments{
\item{x}{A list containing exactly two elements to perform set operations on.}

\item{category_names}{Optional character vector of length 2 specifying the
names to be assigned to the elements of \code{x}. If \code{NULL}, default
names "set1" and "set2" are used.}

\item{max_show}{Maximum number of items to show for the intersection and
union results. If the actual number of items is more than \code{max_show},
only the first \code{max_show} items are displayed.}

\item{max_setdiff_show}{Maximum number of items to show for set difference
results. Similar to \code{max_show}, but applies to set differences.}
}
\value{
A list containing the lengths and actual values (subject to
\code{max_show} and \code{max_setdiff_show}) for each of the set operations
(intersection, union, set difference of set1 from set2, and set difference
of set2 from set1).
}
\description{
This function takes a list of two elements and performs set operations
(intersection, union, and set difference) on these elements. It allows
controlling the output length for intersection, union, and set difference
results.
}
\examples{
set1 <- letters[1:5]
set2 <- letters[3:8]
results <- set_operations(list(set1, set2))
results
}
