% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsv2xlsx.R
\name{tsv2xlsx}
\alias{tsv2xlsx}
\title{Convert TSV file to XLSX (Excel)}
\usage{
tsv2xlsx(file, outfile)
}
\arguments{
\item{file}{A tab-separated (.tsv) file (with header row)}

\item{outfile}{Excel file name}
}
\value{
Nothing
}
\description{
Read TSV file and output it as an XLSX file
}
\examples{
\dontrun{
tsv2xlsx("foo.tsv", "foo.xlsx")
}
}
