% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx2tsv.R
\name{xlsx2tsv}
\alias{xlsx2tsv}
\title{Convert XLSX (Excel) file to TSV}
\usage{
xlsx2tsv(file, outfile, sheet = 1, ...)
}
\arguments{
\item{file}{An Excel file name}

\item{outfile}{A tab-separated (.tsv) file name}

\item{sheet}{The name or index of the Excel sheet to read data from.}

\item{...}{Pass additional params to write_tsv}
}
\value{
Nothing
}
\description{
Read XLSX file and output it as a TSv file
}
\examples{
\dontrun{
xlsx2tsv("foo.xlsx", "foo.tsv")
}
}
