% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colors_polychrome}
\alias{colors_polychrome}
\title{Polychrome color palette (visual impact over colorblind accessibility)}
\format{
A character vector of 25 hex color codes
}
\source{
Created with Polychrome::createPalette(25, c("#FF0000", "#00FF00", "#0000FF"), range = c(30, 80))
}
\usage{
colors_polychrome
}
\description{
Philosophy: Maximizes perceptual distance between colors using algorithmic methods
Uses computational optimization to create colors as visually distinct as possible
Prioritizes visual impact and distinctness over colorblind accessibility
Best for when you need maximum color separation
}
\keyword{datasets}
