% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waldo_compare_files.R
\name{waldo_compare_files}
\alias{waldo_compare_files}
\title{Compare two files}
\usage{
waldo_compare_files(file1, file2, max_diffs = Inf, ...)
}
\arguments{
\item{file1}{Filename for the first file}

\item{file2}{Filename for the second file}

\item{max_diffs}{Maximum no. differences that are reported}

\item{...}{Additional params passed to compare()}
}
\value{
The result of waldo::compare()
}
\description{
Compare two files using waldo::compare
}
\examples{
\dontrun{
waldo_compare_files("foo.txt", "foo.txt")
}
}
